core: remove dead wire-part schemas/exports, consolidate scan+pointer-walk+defineOwn helpers, split conformance memory store; fix wire-v2 remove-then-redeclare + memory-store thread title/atomic projection
