+++ packages/core/src/shape.test.ts
@@ -1,32 +1,36 @@
 import { describe, expect, it } from "vitest";
 import {
-  deriveShape,
   deriveShapeCard,
   describeShape,
-  mergeShapes,
   shapeAtPointer,
   shapeCardSchema,
   type ShapeType,
 } from "./shape.js";
+import type { Json } from "./ids.js";
+
+/** deriveShape/mergeShapes are internal; deriveShapeCard is their only
+ *  production path, so derivation (single sample) and merging (multiple
+ *  samples) are exercised through it. */
+const derived = (...samples: Json[]): ShapeType => deriveShapeCard("t", samples).output;
 
 /** v2 spec §3 — shape cards keep structure only; values are never stored. */
-describe("deriveShape", () => {
+describe("shape derivation (single-sample deriveShapeCard)", () => {
   it("derives scalar kinds", () => {
-    expect(deriveShape("hi")).toEqual({ kind: "string" });
-    expect(deriveShape(3.5)).toEqual({ kind: "number" });
-    expect(deriveShape(true)).toEqual({ kind: "boolean" });
-    expect(deriveShape(null)).toEqual({ kind: "null" });
+    expect(derived("hi")).toEqual({ kind: "string" });
+    expect(derived(3.5)).toEqual({ kind: "number" });
+    expect(derived(true)).toEqual({ kind: "boolean" });
+    expect(derived(null)).toEqual({ kind: "null" });
   });
 
   it("derives objects field-by-field, preserving field order", () => {
-    expect(deriveShape({ month: "Jan", revenue: 1200 })).toEqual({
+    expect(derived({ month: "Jan", revenue: 1200 })).toEqual({
       kind: "object",
       fields: { month: { kind: "string" }, revenue: { kind: "number" } },
     });
   });
 
   it("derives nested structures", () => {
-    expect(deriveShape({ rows: [{ id: 1 }] })).toEqual({
+    expect(derived({ rows: [{ id: 1 }] })).toEqual({
       kind: "object",
       fields: {
         rows: { kind: "array", items: { kind: "object", fields: { id: { kind: "number" } } } },
@@ -35,7 +39,7 @@ describe("deriveShape", () => {
   });
 
   it("merges array element shapes across elements (missing fields become optional)", () => {
-    expect(deriveShape([{ a: 1, b: "x" }, { a: 2 }])).toEqual({
+    expect(derived([{ a: 1, b: "x" }, { a: 2 }])).toEqual({
       kind: "array",
       items: {
         kind: "object",
@@ -46,45 +50,40 @@ describe("deriveShape", () => {
   });
 
   it("derives an empty array's items as json (unknown)", () => {
-    expect(deriveShape([])).toEqual({ kind: "array", items: { kind: "json" } });
+    expect(derived([])).toEqual({ kind: "array", items: { kind: "json" } });
   });
 
   it("degrades non-Json values (undefined, functions) to json instead of throwing", () => {
-    expect(deriveShape(undefined)).toEqual({ kind: "json" });
-    expect(deriveShape(() => 1)).toEqual({ kind: "json" });
+    expect(derived(undefined)).toEqual({ kind: "json" });
+    expect(derived(() => 1)).toEqual({ kind: "json" });
   });
 
   it("caps derivation depth: pathologically deep samples degrade to json, never overflow", () => {
     let deep: unknown = 1;
     for (let i = 0; i < 10_000; i += 1) deep = { next: deep };
-    const shape = deriveShape(deep);
+    const shape = derived(deep);
     expect(shape.kind).toBe("object");
     expect(JSON.stringify(shape)).toContain('"json"');
   });
 });
 
-describe("mergeShapes", () => {
+describe("shape merging (multi-sample deriveShapeCard)", () => {
   it("merges identical scalar kinds to themselves", () => {
-    expect(mergeShapes({ kind: "string" }, { kind: "string" })).toEqual({ kind: "string" });
+    expect(derived("a", "b")).toEqual({ kind: "string" });
   });
 
   it("degrades mismatched kinds to json", () => {
-    expect(mergeShapes({ kind: "string" }, { kind: "number" })).toEqual({ kind: "json" });
-    expect(mergeShapes({ kind: "object", fields: {} }, { kind: "array", items: { kind: "json" } }))
-      .toEqual({ kind: "json" });
+    expect(derived("a", 1)).toEqual({ kind: "json" });
+    expect(derived({}, [])).toEqual({ kind: "json" });
   });
 
   it("json absorbs everything", () => {
-    expect(mergeShapes({ kind: "json" }, { kind: "string" })).toEqual({ kind: "json" });
-    expect(mergeShapes({ kind: "string" }, { kind: "json" })).toEqual({ kind: "json" });
+    expect(derived(undefined, "a")).toEqual({ kind: "json" });
+    expect(derived("a", undefined)).toEqual({ kind: "json" });
   });
 
-  it("merges objects field-wise, marking one-sided fields optional and unioning optional lists", () => {
-    const merged = mergeShapes(
-      { kind: "object", fields: { a: { kind: "number" }, b: { kind: "string" } }, optional: ["b"] },
-      { kind: "object", fields: { a: { kind: "number" }, c: { kind: "boolean" } } },
-    );
-    expect(merged).toEqual({
+  it("merges objects field-wise, marking one-sided fields optional across samples", () => {
+    expect(derived({ a: 1, b: "x" }, { a: 2 }, { a: 3, c: true })).toEqual({
       kind: "object",
       fields: { a: { kind: "number" }, b: { kind: "string" }, c: { kind: "boolean" } },
       optional: ["b", "c"],
@@ -92,10 +91,7 @@ describe("mergeShapes", () => {
   });
 
   it("merges arrays item-wise", () => {
-    expect(mergeShapes(
-      { kind: "array", items: { kind: "number" } },
-      { kind: "array", items: { kind: "number" } },
-    )).toEqual({ kind: "array", items: { kind: "number" } });
+    expect(derived([1], [2])).toEqual({ kind: "array", items: { kind: "number" } });
   });
 });
